// src/config/zod-schema.agent-runtime.d.ts
export declare const ToolsSchema: z.ZodOptional<z.ZodObject<{
// ...
    exec: z.ZodOptional<z.ZodObject<{
        host: z.ZodOptional<z.ZodEnum<{
            auto: "auto";
            sandbox: "sandbox";
            gateway: "gateway";
            node: "node";
        }>>;
// ...
    }, z.core.$strict>>;
// ...
}, z.core.$strict>>;